Added options when ran

+++ main.py
@@ -9,12 +9,30 @@ import user
 def run():
     # Using a try catch to make keyboard interruptions better visually
     try:
-        # Working Email and Password for testing purposes "eve.holt@reqres.in" "cityslicka"
-        # login()
-        # get_users()
-        # get_user()
-        # create_user()
-        user.delete_user()
+        selection = 0
+
+        while selection < 1 and selection > 5:
+            print("Select one of the Following:\n  1. Login\n  2. Get List of Users\n  3. Get Single User\n  4. Create a User\n  5. Delete a User")
+            try:
+                selection = int(input("Please enter the corresponding number: ").strip())
+
+                if selection < 1 and selection > 5:
+                    print("Please enter a valid number between 1-5")
+            except ValueError:
+                print("Please Enter a Number.\n")
+            
+            if selection == 1:
+                # Working Email and Password for testing purposes "eve.holt@reqres.in" "cityslicka"
+                user.login()
+            elif selection == 2:
+                user.get_users()
+            elif selection == 3:
+                user.get_user()
+            elif selection == 4:
+                user.create_user()
+            elif selection == 5:
+                user.delete_user()
+
     except KeyboardInterrupt:
         print('\n\nQuiting!')
         quit()
